+++ ChangeLog.md
@@ -3,7 +3,7 @@ All notable changes to this project will be documented in this file.
 
 ## Unreleased
 
-
+- NEW : change user->right to hasRight() function - *23/11/2023* - 2.3.0  
 
 ## Version 2.2